Allow custom PostgreSQL installation directory.
We already depend on suitable registry entries to find the PostgreSQL versions
for which the extension should be built, so we can also automatically discover
each version root directory at the same time to remove the requirement on
having PostgreSQL installed in the default directory at compile time.

Thanks to Tatsuro Yamada for the feature request.